requirements.txt
numpy
matplotlib
pandas
seaborn
progressbar2
pysam
pyvcf